backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
email-validator
python-dotenv
python-jose
sqlalchemy
minio
alembic
python-multipart
numpy
pillow
ultralytics
pytest
httpx